+++ tests/test_extensions.py
@@ -7,6 +7,7 @@ from test_helpers.utils import ensure_test_package_installed, skip_if_trio
 
 from inspect_ai import Task, eval_async
 from inspect_ai.dataset import Sample
+from inspect_ai.hooks._hooks import emit_run_start
 from inspect_ai.model import get_model
 from inspect_ai.scorer import includes
 from inspect_ai.solver import generate, use_tools
@@ -107,3 +108,13 @@ def test_supports_str_config():
     assert recreated == spec
     assert recreated.config == spec.config
     assert isinstance(recreated.config, str)
+
+
+async def test_hooks():
+    ensure_test_package_installed()
+    module = importlib.import_module("inspect_package.hooks.custom")
+    module.run_ids = []
+
+    await emit_run_start(run_id="42", tasks=[])
+
+    assert module.run_ids == ["42"]